version_2_sequential_website_agent/pyproject.toml
[project]
name = "version-2-sequential-website-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.15.1",
]
